# apps/aurora-core/build/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "aurora-core",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "/api"
		}
	],
	"components": {
...
		"securitySchemes": {
			"local": {
				"type": "apiKey",
				"name": "connect.sid",
				"in": "cookie"
			},
			"integration": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		},
		"schemas": {
...
			"ExternalUrls": {
				"properties": {
					"spotify": {
						"type": "string"
					}
				},
				"required": [
					"spotify"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Followers": {
				"properties": {
					"href": {
						"type": "string",
						"nullable": true
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"href",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Image": {
				"properties": {
					"url": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"url",
					"height",
					"width"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SpotifyUserProfile": {
				"properties": {
					"display_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"external_urls": {
						"$ref": "#/components/schemas/ExternalUrls"
					},
					"followers": {
						"$ref": "#/components/schemas/Followers"
					},
					"href": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/Image"
						},
						"type": "array"
					},
					"type": {
						"type": "string"
					},
					"uri": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"explicit_content": {
						"properties": {
							"filter_locked": {
								"type": "boolean"
							},
							"filter_enabled": {
								"type": "boolean"
							}
						},
						"required": [
							"filter_locked",
							"filter_enabled"
						],
						"type": "object"
					},
					"product": {
						"type": "string"
					}
				},
				"required": [
					"display_name",
					"email",
					"external_urls",
					"followers",
					"href",
					"id",
					"images",
					"type",
					"uri",
					"country",
					"explicit_content",
					"product"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SpotifyUserResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"spotifyId": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"spotifyId",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrackChangeEvent": {
				"properties": {
					"title": {
						"type": "string"
					},
					"artists": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startTime": {
						"type": "string",
						"format": "date-time"
					},
					"cover": {
						"type": "string"
					},
					"trackURI": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"artists",
					"startTime",
					"trackURI"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		}
	},
	"paths": {
...
		"/spotify/login": {
			"get": {
				"operationId": "SpotifyLogin",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Login a user via Spotify. Used to bind the returned spotify user to this server instance.\nPerforms a redirect to the Spotify website",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/spotify/callback": {
			"get": {
				"operationId": "SpotifyLoginCallback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SpotifyUserProfile"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"412": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "state",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "error",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/spotify/user/current": {
			"get": {
				"operationId": "GetCurrentSpotifyUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SpotifyUserResponse"
								}
							}
						}
					},
					"204": {
						"description": "No user active"
					}
				},
				"description": "Get the currently active Spotify user.",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
		"/spotify/users": {
			"get": {
				"operationId": "GetAllSpotifyUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SpotifyUserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all existing Spotify users",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/spotify/users/{id}": {
			"delete": {
				"operationId": "DeleteSpotifyUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete an existing Spotify user",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/spotify/users/{id}/switch": {
			"post": {
				"operationId": "SwitchToSpotifyUser",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Switch the currently active Spotify user to the user with the given ID",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/spotify/profile": {
			"get": {
				"operationId": "GetSpotifyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SpotifyUserProfile"
								}
							}
						}
					},
					"204": {
						"description": "No user active"
					}
				},
				"description": "Get the Spotify profile from the currently active user",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
		"/spotify/currently-playing": {
			"get": {
				"operationId": "GetSpotifyCurrentlyPlaying",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrackChangeEvent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get the currently playing track (on Spotify or locally playing). Null if nothing is playing.",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
		"/spotify/skip": {
			"post": {
				"operationId": "SkipSpotifyTrack",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Skip the currently playing Spotify track.",
				"tags": [
					"Spotify"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
...
	}
}